+++ CHANGELOG.md
@@ -4,5 +4,8 @@
 <!-- fix: A bug fix -->
 <!-- chore: Changes to the build process or auxiliary tools and libraries -->
 
+### 0.1.1
+- fix: Fix for issue with registered lifecycle events. 
+
 ### 0.1.0
-- add: Initial release of the Sprig Flutter plugin.
+- add: Initial Sprig Flutter plugin.